Finally finished what I set out to do.  I'm going to sleep now
clientserver and embedded are finished for everything except RequestTable (to my current knowledge)
