+++ src/main/java/edu/wpi/GoldenGandaberundas/tableControllers/AudioVisualService/AudioVisualRequestClientServer.java
@@ -1,7 +1,8 @@
-package edu.wpi.GoldenGandaberundas.tableControllers.GiftDeliveryService;
+package edu.wpi.GoldenGandaberundas.tableControllers.AudioVisualService;
 
 import edu.wpi.GoldenGandaberundas.TableController;
 import edu.wpi.GoldenGandaberundas.tableControllers.DBConnection.ConnectionHandler;
+import edu.wpi.GoldenGandaberundas.tableControllers.AudioVisualService.AudioVisualRequest;
 import edu.wpi.GoldenGandaberundas.tableControllers.Requests.Request;
 import edu.wpi.GoldenGandaberundas.tableControllers.Requests.RequestTable;
 
@@ -14,8 +15,8 @@ import java.util.List;
 import java.util.Locale;
 import java.util.stream.Collectors;
 
-public class GiftRequestClientServer
-        implements TableController<GiftRequest, ArrayList<Integer>> {
+public class AudioVisualRequestClientServer
+        implements TableController<AudioVisualRequest, ArrayList<Integer>> {
     private static TableController<Request, Integer> masterTable = null;
     /** name of table */
     private String tbName;
@@ -24,14 +25,14 @@ public class GiftRequestClientServer
     /** list of keys that make a composite primary key */
     private String pkCols = null;
     /** list that contains the objects stored in the database */
-    private ArrayList<GiftRequest> objList;
+    private ArrayList<AudioVisualRequest> objList;
     /** relative path to the database file */
     ConnectionHandler connectionHandler = ConnectionHandler.getInstance();
 
     Connection connection = connectionHandler.getConnection();
 
-    public GiftRequestClientServer(
-            String tbName, String[] cols, String pkCols, ArrayList<GiftRequest> objList)
+    public AudioVisualRequestClientServer(
+            String tbName, String[] cols, String pkCols, ArrayList<AudioVisualRequest> objList)
             throws SQLException {
         // create a new table with column names if none table of same name exist
         // if there is one, do nothing
@@ -42,26 +43,25 @@ public class GiftRequestClientServer
     }
 
     @Override
-    public ArrayList<GiftRequest> readTable() {
-        ArrayList tableInfo = new ArrayList<GiftRequest>(); // **
+    public ArrayList<AudioVisualRequest> readTable() {
+        ArrayList tableInfo = new ArrayList<AudioVisualRequest>(); // **
         try {
             PreparedStatement s = connection.prepareStatement("SElECT * FROM " + tbName + ";");
             ResultSet r = s.executeQuery();
             while (r.next()) {
-                System.out.println(masterTable.getEntry(r.getInt(1)));
                 tableInfo.add(
-                        new GiftRequest( // **
-                                masterTable.getEntry(r.getInt(1)), r.getInt(2), r.getInt(3)));
+                        new AudioVisualRequest(masterTable.getEntry(r.getInt(1)), r.getInt(2), r.getString(3)));
             }
         } catch (SQLException se) {
             se.printStackTrace();
             return null;
         }
+        objList = tableInfo;
         return tableInfo;
     }
 
     @Override
-    public boolean addEntry(GiftRequest gift) {
+    public boolean addEntry(AudioVisualRequest audioVisual) {
         try {
             PreparedStatement s =
                     connection.prepareStatement(
@@ -80,11 +80,11 @@ public class GiftRequestClientServer
                                     + " VALUES (?, ?, ?)"
                                     + "end");
 
-            s.setInt(1, gift.getRequestID());
-            s.setInt(2, gift.getGiftID());
-            s.setInt(3, gift.getRequestID());
-            s.setInt(4, gift.getGiftID());
-            s.setInt(5, gift.getQuantity());
+            s.setInt(1, audioVisual.getRequestID());
+            s.setInt(2, audioVisual.getAudioVisualID());
+            s.setInt(3, audioVisual.getRequestID());
+            s.setInt(4, audioVisual.getAudioVisualID());
+            s.setString(5, audioVisual.getPriority());
             s.executeUpdate();
             return true;
         } catch (SQLException e) {
@@ -94,21 +94,26 @@ public class GiftRequestClientServer
     }
 
     @Override
-    public ArrayList<GiftRequest> readBackup(String fileName) {
-        ArrayList<GiftRequest> gReqList = new ArrayList<GiftRequest>(); // **
+    public ArrayList<AudioVisualRequest> readBackup(String fileName) {
+        ArrayList<AudioVisualRequest> avReqList = new ArrayList<>(); // **
 
         try {
             File csvFile = new File(fileName);
             BufferedReader buffer = new BufferedReader(new FileReader(csvFile)); // reads the files
             String currentLine = buffer.readLine(); // reads a line from the csv file
             System.out.println(currentLine);
-
+            if (!currentLine
+                    .toLowerCase(Locale.ROOT)
+                    .trim()
+                    .equals(new String("reqID,audioVisualID,priority"))) { // **
+                System.err.println("audio visual request backup format not recognized"); // **
+            }
             currentLine = buffer.readLine();
 
             while (currentLine != null) { // cycles in the while loop until it reaches the end
                 String[] element = currentLine.split(","); // separates each element based on a comma
-                GiftRequest req = // **
-                        new GiftRequest(
+                AudioVisualRequest req = // **
+                        new AudioVisualRequest(
                                 Integer.parseInt(element[0]),
                                 element[1],
                                 Integer.parseInt(element[2]),
@@ -119,8 +124,8 @@ public class GiftRequestClientServer
                                 element[8],
                                 element[9],
                                 Integer.parseInt(element[10]),
-                                Integer.parseInt(element[11]));
-                gReqList.add(req); // adds the location to the list
+                                element[11]);
+                avReqList.add(req); // adds the location to the list
                 currentLine = buffer.readLine();
             }
             ; // creates a Location
@@ -130,7 +135,7 @@ public class GiftRequestClientServer
         } catch (IOException ex) {
             ex.printStackTrace();
         }
-        return gReqList; // **
+        return avReqList; // **
     }
 
     @Override
@@ -147,15 +152,15 @@ public class GiftRequestClientServer
             }
             Statement s = connection.createStatement();
             s.execute(
-                    "CREATE TABLE  GiftRequests("
+                    "CREATE TABLE  AudioVisualRequests("
                             + "reqID INTEGER NOT NULL, "
-                            + "giftID INTEGER NOT NULL, "
-                            + "quantity INTEGER, "
-                            + "CONSTRAINT GiftRequestPK PRIMARY KEY (reqID, giftID), "
+                            + "audioVisualID INTEGER NOT NULL, "
+                            + "priority TEXT NOT NULL, "
+                            + "CONSTRAINT AudioVisualRequestPK PRIMARY KEY (reqID, audioVisualID), "
                             + "CONSTRAINT RequestFK FOREIGN KEY (reqID) REFERENCES Requests (requestID) "
                             + "ON UPDATE CASCADE "
                             + "ON DELETE CASCADE, "
-                            + "CONSTRAINT GiftFK FOREIGN KEY (giftID) REFERENCES Gifts (giftID) "
+                            + "CONSTRAINT AudioVisualFK FOREIGN KEY (audioVisualID) REFERENCES AudioVisual (avID) "
                             + "ON UPDATE CASCADE "
                             + "ON DELETE CASCADE);");
         } catch (SQLException e) {
@@ -164,8 +169,8 @@ public class GiftRequestClientServer
     }
 
     @Override
-    public GiftRequest getEntry(ArrayList<Integer> pkID) {
-        GiftRequest gift = new GiftRequest();
+    public AudioVisualRequest getEntry(ArrayList<Integer> pkID) {
+        AudioVisualRequest audReq = new AudioVisualRequest();
         if (this.entryExists(pkID)) {
             try {
                 PreparedStatement s =
@@ -177,24 +182,24 @@ public class GiftRequestClientServer
                 r.next();
 
                 if (entryExists(pkID)) {
-                    gift.setRequestID(r.getInt(1));
-                    gift.setGiftID(r.getInt(2));
-                    gift.setQuantity(r.getInt(3));
+                    audReq.setRequestID(r.getInt(1));
+                    audReq.setAudioVisualID(r.getInt(2));
+                    audReq.setPriority(r.getString(3));
                 }
-                return gift; // **
+                return audReq; // **
             } catch (SQLException e) {
                 e.printStackTrace();
             }
         }
-        return gift; // **
+        return audReq; // **
     }
 
     @Override
-    public boolean loadFromArrayList(ArrayList<GiftRequest> objList) {
+    public boolean loadFromArrayList(ArrayList<AudioVisualRequest> objList) {
         this.createTable();
         deleteTableData();
-        for (GiftRequest giftRequest : objList) {
-            if (!this.addEntry(giftRequest)) {
+        for (AudioVisualRequest audioVisualRequest : objList) {
+            if (!this.addEntry(audioVisualRequest)) {
                 return false;
             }
         }
@@ -213,7 +218,7 @@ public class GiftRequestClientServer
     @Override
     public void writeTable() {
 
-        for (GiftRequest obj : objList) {
+        for (AudioVisualRequest obj : objList) {
 
             this.addEntry(obj);
         }
@@ -337,9 +342,9 @@ public class GiftRequestClientServer
     }
 
     // drop current table and enter data from CSV
-    public ArrayList<GiftRequest> loadBackup(String fileName) {
+    public ArrayList<AudioVisualRequest> loadBackup(String fileName) {
         createTable();
-        ArrayList<GiftRequest> listObjs = readBackup(fileName);
+        ArrayList<AudioVisualRequest> listObjs = readBackup(fileName);
 
         try {
             PreparedStatement s = connection.prepareStatement("DELETE FROM " + tbName + ";");
@@ -387,7 +392,7 @@ public class GiftRequestClientServer
         return tbName;
     }
 
-    public ArrayList<GiftRequest> getObjList() {
+    public ArrayList<AudioVisualRequest> getObjList() {
         return objList;
     }
 }
