+++ src/main/java/edu/wpi/GoldenGandaberundas/tableControllers/AudioVisualService/AudioVisualRequestTbl.java
@@ -1,7 +1,11 @@
 package edu.wpi.GoldenGandaberundas.tableControllers.AudioVisualService;
 
 import edu.wpi.GoldenGandaberundas.TableController;
+import edu.wpi.GoldenGandaberundas.tableControllers.AudioVisualService.AudioVisualRequest;
+import edu.wpi.GoldenGandaberundas.tableControllers.AudioVisualService.AudioVisualRequestClientServer;
+import edu.wpi.GoldenGandaberundas.tableControllers.AudioVisualService.AudioVisualRequestEmbedded;
 import edu.wpi.GoldenGandaberundas.tableControllers.DBConnection.ConnectionHandler;
+import edu.wpi.GoldenGandaberundas.tableControllers.DBConnection.ConnectionType;
 import edu.wpi.GoldenGandaberundas.tableControllers.Requests.Request;
 import edu.wpi.GoldenGandaberundas.tableControllers.Requests.RequestTable;
 import java.io.*;
@@ -28,6 +32,10 @@ public class AudioVisualRequestTbl
   /** list that contains the objects stored in the database */
   protected ArrayList<AudioVisualRequest> objList;
   /** relative path to the database file */
+  TableController<AudioVisualRequest, ArrayList<Integer>> embeddedTable = null;
+
+  TableController<AudioVisualRequest, ArrayList<Integer>> clientServerTable = null;
+
   ConnectionHandler connectionHandler = ConnectionHandler.getInstance();
 
   Connection connection = connectionHandler.getConnection();
@@ -40,8 +48,15 @@ public class AudioVisualRequestTbl
     pkCols = "reqID, audioVisualID";
     String[] cols = {"reqID", "audioVisualID", "priority"};
     masterTable = RequestTable.getInstance();
-    createTable();
     objList = new ArrayList<AudioVisualRequest>();
+    embeddedTable =
+            new AudioVisualRequestEmbedded(tbName, colNames.toArray(new String[3]), pkCols, objList);
+    clientServerTable =
+            new AudioVisualRequestClientServer(tbName, colNames.toArray(new String[3]), pkCols, objList);
+    connectionHandler.addTable(embeddedTable, ConnectionType.embedded);
+    connectionHandler.addTable(clientServerTable, ConnectionType.clientServer);
+    masterTable = RequestTable.getInstance();
+    createTable();
     objList = readTable();
   }
 
@@ -62,178 +77,43 @@ public class AudioVisualRequestTbl
     return instance;
   }
 
+  private TableController<AudioVisualRequest, ArrayList<Integer>> getCurrentTable() {
+    System.out.println("Connection Type: " + connectionHandler.getCurrentConnectionType());
+    switch (connectionHandler.getCurrentConnectionType()) {
+      case embedded:
+        return embeddedTable;
+      case clientServer:
+        return clientServerTable;
+      case cloud:
+        return null;
+    }
+    System.out.println(connectionHandler.getCurrentConnectionType());
+    return null;
+  }
+
   // Creates readTable method and returns an array list of AudioVisual Requests
-  @Override
-  public ArrayList<AudioVisualRequest> readTable() {
-    ArrayList tableInfo = new ArrayList<AudioVisualRequest>(); // **
-    try {
-      PreparedStatement s = connection.prepareStatement("SElECT * FROM " + tbName + ";");
-      ResultSet r = s.executeQuery();
-      while (r.next()) {
-        tableInfo.add(
-            new AudioVisualRequest(masterTable.getEntry(r.getInt(1)), r.getInt(2), r.getString(3)));
-      }
-    } catch (SQLException se) {
-      se.printStackTrace();
-      return null;
-    }
-    objList = tableInfo;
-    return tableInfo;
-  }
 
   @Override
-  public boolean addEntry(AudioVisualRequest obj) {
-    if (!RequestTable.getInstance().entryExists(obj.getRequestID())) {
-      RequestTable.getInstance().addEntry(obj);
-    }
-    AudioVisualRequest audioVisualReq = (AudioVisualRequest) obj; // **
-    PreparedStatement s = null;
-    try {
-      s =
-          connection.prepareStatement( // **
-              "INSERT OR IGNORE INTO " + tbName + " VALUES (?, ?, ?);");
-
-      s.setInt(1, obj.getRequestID());
-      s.setInt(2, obj.getAudioVisualID());
-      s.setString(3, obj.getPriority());
-      s.executeUpdate();
-      return true;
-    } catch (SQLException e) {
-      e.printStackTrace();
-      return false;
-    }
-  }
+  public ArrayList<AudioVisualRequest> readTable() {return this.getCurrentTable().readTable();}
 
   @Override
-  public ArrayList<AudioVisualRequest> readBackup(String fileName) {
-    ArrayList<AudioVisualRequest> avReqList = new ArrayList<>(); // **
-
-    try {
-      File csvFile = new File(fileName);
-      BufferedReader buffer = new BufferedReader(new FileReader(csvFile)); // reads the files
-      String currentLine = buffer.readLine(); // reads a line from the csv file
-      System.out.println(currentLine);
-      if (!currentLine
-          .toLowerCase(Locale.ROOT)
-          .trim()
-          .equals(new String("reqID,audioVisualID,priority"))) { // **
-        System.err.println("audio visual request backup format not recognized"); // **
-      }
-      currentLine = buffer.readLine();
-
-      while (currentLine != null) { // cycles in the while loop until it reaches the end
-        String[] element = currentLine.split(","); // separates each element based on a comma
-        AudioVisualRequest req = // **
-            new AudioVisualRequest(
-                Integer.parseInt(element[0]),
-                element[1],
-                Integer.parseInt(element[2]),
-                Integer.parseInt(element[3]),
-                Long.parseLong(element[4]),
-                Long.parseLong(element[5]),
-                Integer.parseInt(element[6]),
-                element[8],
-                element[9],
-                Integer.parseInt(element[10]),
-                element[11]);
-        avReqList.add(req); // adds the location to the list
-        currentLine = buffer.readLine();
-      }
-      ; // creates a Location
-
-    } catch (FileNotFoundException ex) {
-      ex.printStackTrace();
-    } catch (IOException ex) {
-      ex.printStackTrace();
-    }
-    return avReqList; // **
-  }
+  public boolean addEntry(AudioVisualRequest obj) {return this.getCurrentTable().addEntry(obj);}
 
   @Override
-  public void createTable() {
-    try {
-      PreparedStatement s =
-          connection.prepareStatement("SELECT count(*) FROM sqlite_master WHERE tbl_name = ?;");
-      s.setString(1, tbName);
-      ResultSet r = s.executeQuery();
-      r.next();
-      if (r.getInt(1) != 0) {
-        return;
-      }
-    } catch (SQLException e) {
-      e.printStackTrace();
-      return;
-    }
-
-    try {
-      Class.forName("org.sqlite.JDBC");
-    } catch (ClassNotFoundException e) {
-      System.out.println("SQLite driver not found on classpath, check your gradle configuration.");
-      e.printStackTrace();
-      return;
-    }
-
-    System.out.println("SQLite driver registered!");
-
-    Statement s = null;
-    try {
-      s = connection.createStatement();
-      s.execute("PRAGMA foreign_keys = ON"); // **
-      s.execute(
-          "CREATE TABLE  AudioVisualRequests("
-              + "reqID INTEGER NOT NULL, "
-              + "audioVisualID INTEGER NOT NULL, "
-              + "priority TEXT NOT NULL, "
-              + "CONSTRAINT AudioVisualRequestPK PRIMARY KEY (reqID, audioVisualID), "
-              + "CONSTRAINT RequestFK FOREIGN KEY (reqID) REFERENCES Requests (requestID) "
-              + "ON UPDATE CASCADE "
-              + "ON DELETE CASCADE, "
-              + "CONSTRAINT AudioVisualFK FOREIGN KEY (audioVisualID) REFERENCES AudioVisual (avID) "
-              + "ON UPDATE CASCADE "
-              + "ON DELETE CASCADE);");
-
-    } catch (SQLException e) {
-      e.printStackTrace();
-    }
-  }
+  public ArrayList<AudioVisualRequest> readBackup(String fileName) {return this.getCurrentTable().readBackup(fileName);}
 
   @Override
-  public AudioVisualRequest getEntry(ArrayList<Integer> pkID) {
-    AudioVisualRequest audReq = new AudioVisualRequest();
-    if (this.entryExists(pkID)) {
-      try {
-        PreparedStatement s =
-            connection.prepareStatement(
-                "SELECT * FROM " + tbName + " WHERE (" + pkCols + ") =(?,?);");
-        s.setInt(1, pkID.get(0));
-        s.setInt(2, pkID.get(1));
-        ResultSet r = s.executeQuery();
-        r.next();
+  public void createTable() {this.getCurrentTable().createTable();}
 
-        if (entryExists(pkID)) {
-          audReq.setRequestID(r.getInt(1));
-          audReq.setAudioVisualID(r.getInt(2));
-          audReq.setPriority(r.getString(3));
-        }
-        return audReq; // **
-      } catch (SQLException e) {
-        e.printStackTrace();
-      }
-    }
-    return audReq; // **
-  }
+  @Override
+  public AudioVisualRequest getEntry(ArrayList<Integer> pkID) {return this.getCurrentTable().getEntry(pkID);}
 
   @Override
   public boolean loadFromArrayList(ArrayList<AudioVisualRequest> objList) {
-    return false;
+    return this.getCurrentTable().loadFromArrayList(objList);
   }
 
-  public void writeTable() {
-
-    for (AudioVisualRequest obj : objList) {
-      this.addEntry(obj);
-    }
-  }
+  public void writeTable() {this.getCurrentTable().writeTable();}
 
   /**
    * Modifies the attribute so that it is equal to value MAKE SURE YOU KNOW WHAT DATA TYPE YOU ARE
@@ -245,33 +125,8 @@ public class AudioVisualRequestTbl
    * @return true if successful, false otherwise
    */
   // public boolean editEntry(T1 pkid, String colName, Object value)
-  public boolean editEntry(ArrayList<Integer> pkid, String colName, Object value) {
-    StringBuilder pkString = new StringBuilder();
-    for (int i = 0; i < pkid.size() - 1; i++) {
-      pkString.append(pkid.get(i)).append(",");
-    }
-    pkString.append(pkid.get(pkid.size() - 1));
-    try {
+  public boolean editEntry(ArrayList<Integer> pkid, String colName, Object value) {return this.getCurrentTable().editEntry(pkid, colName, value);}
 
-      PreparedStatement s =
-          connection.prepareStatement(
-              "UPDATE "
-                  + tbName
-                  + " SET "
-                  + colName
-                  + " = ? WHERE ("
-                  + pkCols
-                  + ") = ("
-                  + pkString
-                  + ");");
-      s.setObject(1, value);
-      s.executeUpdate();
-      return true;
-    } catch (SQLException e) {
-      e.printStackTrace();
-      return false;
-    }
-  }
 
   /**
    * removes a row from the database
@@ -279,125 +134,21 @@ public class AudioVisualRequestTbl
    * @param pkid primary key of row to be removed
    * @return true if successful, false otherwise
    */
-  public boolean deleteEntry(ArrayList<Integer> pkid) {
-    StringBuilder pkString = new StringBuilder();
-    for (int i = 0; i < pkid.size() - 1; i++) {
-      pkString.append(pkid.get(i)).append(",");
-    }
-    pkString.append(pkid.get(pkid.size() - 1));
-    try {
-      PreparedStatement s =
-          connection.prepareStatement(
-              "DELETE FROM " + tbName + " WHERE (" + pkCols + ") = (" + pkString + ");");
-      s.executeUpdate();
-      return true;
-    } catch (SQLException e) {
-      e.printStackTrace();
-    }
-    return false;
-  }
+  public boolean deleteEntry(ArrayList<Integer> pkid) {return this.getCurrentTable().deleteEntry(pkid);}
+
 
   /**
    * creates CSV file representing the objects stored in the table
    *
    * @param f filename of the to be created CSV
    */
-  public void createBackup(File f) {
-    if (objList.isEmpty()) {
-      return;
-    }
-    /* Instantiate the writer */
-    PrintWriter writer = null;
-    try {
-      writer = new PrintWriter(f);
-    } catch (FileNotFoundException e) {
-      e.printStackTrace();
-    }
-
-    /* Get the class type of the objects in the array */
-    final Class<?> type = objList.get(0).getClass();
-
-    /* Get the name of all the attributes */
-    final ArrayList<Field> classAttributes = new ArrayList<>(List.of(type.getDeclaredFields()));
-
-    boolean doesExtend = Request.class.isAssignableFrom(type);
-    if (doesExtend) {
-      final Class<?> superType = objList.get(0).getClass().getSuperclass();
-      classAttributes.addAll(0, (List.of(superType.getDeclaredFields())));
-    }
-
-    /* Write the parsed attributes to the file */
-    writer.println(classAttributes.stream().map(Field::getName).collect(Collectors.joining(",")));
-
-    /* For each object, read each attribute and append it to the file with a comma separating */
-    PrintWriter finalWriter = writer;
-    objList.forEach(
-        obj -> {
-          finalWriter.println(
-              classAttributes.stream()
-                  .map(
-                      attribute -> {
-                        attribute.setAccessible(true);
-                        String output = "";
-                        try {
-                          output = attribute.get(obj).toString();
-                        } catch (IllegalAccessException | ClassCastException e) {
-                          System.err.println("[CSVUtil] Object attribute access error.");
-                        }
-                        return output;
-                      })
-                  .collect(Collectors.joining(",")));
-          finalWriter.flush();
-        });
-    writer.close();
-  }
+  public void createBackup(File f) {this.getCurrentTable().createBackup(f);}
 
   // drop current table and enter data from CSV
-  public ArrayList<AudioVisualRequest> loadBackup(String fileName) {
-    createTable();
-    ArrayList<AudioVisualRequest> listObjs = readBackup(fileName);
-
-    try {
-      PreparedStatement s = connection.prepareStatement("DELETE FROM " + tbName + ";");
-      s.executeUpdate();
-      this.objList = listObjs;
-      this.writeTable();
-    } catch (SQLException e) {
-      e.printStackTrace();
-    }
-    return listObjs;
-  }
+  public ArrayList<AudioVisualRequest> loadBackup(String fileName) {return this.getCurrentTable().loadBackup(fileName);}
 
   // checks if an entry exists
-  public boolean entryExists(ArrayList<Integer> pkID) {
-    boolean exists = false;
-    StringBuilder pkString = new StringBuilder();
-    for (int i = 0; i < pkID.size() - 1; i++) {
-      pkString.append(pkID.get(i)).append(",");
-    }
-    pkString.append(pkID.get(pkID.size() - 1));
-
-    try {
-      PreparedStatement s =
-          connection.prepareStatement(
-              "SELECT count(*) FROM "
-                  + tbName
-                  + " WHERE ("
-                  + pkCols
-                  + ") = ("
-                  + pkString.toString()
-                  + ");");
-      ResultSet r = s.executeQuery();
-      r.next();
-      if (r.getInt(1) != 0) {
-        exists = true;
-      }
-      return exists;
-    } catch (SQLException e) {
-      e.printStackTrace();
-      return false;
-    }
-  }
+  public boolean entryExists(ArrayList<Integer> pkID) {return this.getCurrentTable().entryExists(pkID);}
 
   public String getTableName() {
     return tbName;
