+++ src/main/java/edu/wpi/GoldenGandaberundas/tableControllers/FoodService/FoodRequestClientServer.java
@@ -1,4 +1,4 @@
-package edu.wpi.GoldenGandaberundas.tableControllers.GiftDeliveryService;
+package edu.wpi.GoldenGandaberundas.tableControllers.FoodService;
 
 import edu.wpi.GoldenGandaberundas.TableController;
 import edu.wpi.GoldenGandaberundas.tableControllers.DBConnection.ConnectionHandler;
@@ -14,8 +14,8 @@ import java.util.List;
 import java.util.Locale;
 import java.util.stream.Collectors;
 
-public class GiftRequestClientServer
-        implements TableController<GiftRequest, ArrayList<Integer>> {
+public class FoodRequestClientServer
+        implements TableController<FoodRequest, ArrayList<Integer>> {
     private static TableController<Request, Integer> masterTable = null;
     /** name of table */
     private String tbName;
@@ -24,14 +24,14 @@ public class GiftRequestClientServer
     /** list of keys that make a composite primary key */
     private String pkCols = null;
     /** list that contains the objects stored in the database */
-    private ArrayList<GiftRequest> objList;
+    private ArrayList<FoodRequest> objList;
     /** relative path to the database file */
     ConnectionHandler connectionHandler = ConnectionHandler.getInstance();
 
     Connection connection = connectionHandler.getConnection();
 
-    public GiftRequestClientServer(
-            String tbName, String[] cols, String pkCols, ArrayList<GiftRequest> objList)
+    public FoodRequestClientServer(
+            String tbName, String[] cols, String pkCols, ArrayList<FoodRequest> objList)
             throws SQLException {
         // create a new table with column names if none table of same name exist
         // if there is one, do nothing
@@ -42,26 +42,24 @@ public class GiftRequestClientServer
     }
 
     @Override
-    public ArrayList<GiftRequest> readTable() {
-        ArrayList tableInfo = new ArrayList<GiftRequest>(); // **
+    public ArrayList<FoodRequest> readTable() {
+        ArrayList tableInfo = new ArrayList<FoodRequest>(); // **
         try {
             PreparedStatement s = connection.prepareStatement("SElECT * FROM " + tbName + ";");
             ResultSet r = s.executeQuery();
             while (r.next()) {
-                System.out.println(masterTable.getEntry(r.getInt(1)));
-                tableInfo.add(
-                        new GiftRequest( // **
-                                masterTable.getEntry(r.getInt(1)), r.getInt(2), r.getInt(3)));
+                tableInfo.add(new FoodRequest(masterTable.getEntry(r.getInt(1)), r.getInt(2), r.getInt(3)));
             }
         } catch (SQLException se) {
             se.printStackTrace();
             return null;
         }
+        objList = tableInfo;
         return tableInfo;
     }
 
     @Override
-    public boolean addEntry(GiftRequest gift) {
+    public boolean addEntry(FoodRequest food) {
         try {
             PreparedStatement s =
                     connection.prepareStatement(
@@ -80,11 +78,11 @@ public class GiftRequestClientServer
                                     + " VALUES (?, ?, ?)"
                                     + "end");
 
-            s.setInt(1, gift.getRequestID());
-            s.setInt(2, gift.getGiftID());
-            s.setInt(3, gift.getRequestID());
-            s.setInt(4, gift.getGiftID());
-            s.setInt(5, gift.getQuantity());
+            s.setInt(1, food.getRequestID());
+            s.setInt(2, food.getFoodID());
+            s.setInt(3, food.getRequestID());
+            s.setInt(4, food.getFoodID());
+            s.setInt(5, food.getQuantity());
             s.executeUpdate();
             return true;
         } catch (SQLException e) {
@@ -94,21 +92,27 @@ public class GiftRequestClientServer
     }
 
     @Override
-    public ArrayList<GiftRequest> readBackup(String fileName) {
-        ArrayList<GiftRequest> gReqList = new ArrayList<GiftRequest>(); // **
+    public ArrayList<FoodRequest> readBackup(String fileName) {
+        ArrayList<FoodRequest> fReqList = new ArrayList<>(); // **
 
         try {
             File csvFile = new File(fileName);
             BufferedReader buffer = new BufferedReader(new FileReader(csvFile)); // reads the files
             String currentLine = buffer.readLine(); // reads a line from the csv file
             System.out.println(currentLine);
+            if (!currentLine
+                    .toLowerCase(Locale.ROOT)
+                    .trim()
+                    .equals(new String("reqID,foodID,quantity"))) { // **
+                System.err.println("food request backup format not recognized"); // **
+            }
 
             currentLine = buffer.readLine();
 
             while (currentLine != null) { // cycles in the while loop until it reaches the end
                 String[] element = currentLine.split(","); // separates each element based on a comma
-                GiftRequest req = // **
-                        new GiftRequest(
+                FoodRequest req = // **
+                        new FoodRequest(
                                 Integer.parseInt(element[0]),
                                 element[1],
                                 Integer.parseInt(element[2]),
@@ -120,7 +124,7 @@ public class GiftRequestClientServer
                                 element[9],
                                 Integer.parseInt(element[10]),
                                 Integer.parseInt(element[11]));
-                gReqList.add(req); // adds the location to the list
+                fReqList.add(req); // adds the location to the list
                 currentLine = buffer.readLine();
             }
             ; // creates a Location
@@ -130,7 +134,7 @@ public class GiftRequestClientServer
         } catch (IOException ex) {
             ex.printStackTrace();
         }
-        return gReqList; // **
+        return fReqList; // **
     }
 
     @Override
@@ -147,15 +151,15 @@ public class GiftRequestClientServer
             }
             Statement s = connection.createStatement();
             s.execute(
-                    "CREATE TABLE  GiftRequests("
+                    "CREATE TABLE  FoodRequests("
                             + "reqID INTEGER NOT NULL, "
-                            + "giftID INTEGER NOT NULL, "
+                            + "foodID INTEGER NOT NULL, "
                             + "quantity INTEGER, "
-                            + "CONSTRAINT GiftRequestPK PRIMARY KEY (reqID, giftID), "
+                            + "CONSTRAINT FoodRequestPK PRIMARY KEY (reqID, foodID), "
                             + "CONSTRAINT RequestFK FOREIGN KEY (reqID) REFERENCES Requests (requestID) "
                             + "ON UPDATE CASCADE "
                             + "ON DELETE CASCADE, "
-                            + "CONSTRAINT GiftFK FOREIGN KEY (giftID) REFERENCES Gifts (giftID) "
+                            + "CONSTRAINT FoodFK FOREIGN KEY (foodID) REFERENCES Foods (foodID) "
                             + "ON UPDATE CASCADE "
                             + "ON DELETE CASCADE);");
         } catch (SQLException e) {
@@ -164,8 +168,8 @@ public class GiftRequestClientServer
     }
 
     @Override
-    public GiftRequest getEntry(ArrayList<Integer> pkID) {
-        GiftRequest gift = new GiftRequest();
+    public FoodRequest getEntry(ArrayList<Integer> pkID) {
+        FoodRequest food = new FoodRequest();
         if (this.entryExists(pkID)) {
             try {
                 PreparedStatement s =
@@ -177,24 +181,24 @@ public class GiftRequestClientServer
                 r.next();
 
                 if (entryExists(pkID)) {
-                    gift.setRequestID(r.getInt(1));
-                    gift.setGiftID(r.getInt(2));
-                    gift.setQuantity(r.getInt(3));
+                    food.setRequestID(r.getInt(1));
+                    food.setFoodID(r.getInt(2));
+                    food.setQuantity(r.getInt(3));
                 }
-                return gift; // **
+                return food; // **
             } catch (SQLException e) {
                 e.printStackTrace();
             }
         }
-        return gift; // **
+        return food; // **
     }
 
     @Override
-    public boolean loadFromArrayList(ArrayList<GiftRequest> objList) {
+    public boolean loadFromArrayList(ArrayList<FoodRequest> objList) {
         this.createTable();
         deleteTableData();
-        for (GiftRequest giftRequest : objList) {
-            if (!this.addEntry(giftRequest)) {
+        for (FoodRequest foodRequest : objList) {
+            if (!this.addEntry(foodRequest)) {
                 return false;
             }
         }
@@ -213,7 +217,7 @@ public class GiftRequestClientServer
     @Override
     public void writeTable() {
 
-        for (GiftRequest obj : objList) {
+        for (FoodRequest obj : objList) {
 
             this.addEntry(obj);
         }
@@ -337,9 +341,9 @@ public class GiftRequestClientServer
     }
 
     // drop current table and enter data from CSV
-    public ArrayList<GiftRequest> loadBackup(String fileName) {
+    public ArrayList<FoodRequest> loadBackup(String fileName) {
         createTable();
-        ArrayList<GiftRequest> listObjs = readBackup(fileName);
+        ArrayList<FoodRequest> listObjs = readBackup(fileName);
 
         try {
             PreparedStatement s = connection.prepareStatement("DELETE FROM " + tbName + ";");
@@ -387,7 +391,7 @@ public class GiftRequestClientServer
         return tbName;
     }
 
-    public ArrayList<GiftRequest> getObjList() {
+    public ArrayList<FoodRequest> getObjList() {
         return objList;
     }
 }
