+++ src/main/java/edu/wpi/GoldenGandaberundas/tableControllers/FoodService/FoodRequestTbl.java
@@ -2,6 +2,7 @@ package edu.wpi.GoldenGandaberundas.tableControllers.FoodService;
 
 import edu.wpi.GoldenGandaberundas.TableController;
 import edu.wpi.GoldenGandaberundas.tableControllers.DBConnection.ConnectionHandler;
+import edu.wpi.GoldenGandaberundas.tableControllers.DBConnection.ConnectionType;
 import edu.wpi.GoldenGandaberundas.tableControllers.Requests.Request;
 import edu.wpi.GoldenGandaberundas.tableControllers.Requests.RequestTable;
 import java.io.*;
@@ -10,14 +11,13 @@ import java.sql.*;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Locale;
 import java.util.stream.Collectors;
 
 public class FoodRequestTbl implements TableController<FoodRequest, ArrayList<Integer>> {
 
   // **
   // created instance for singleton
-  private static FoodRequestTbl instance = null;
+  private static edu.wpi.GoldenGandaberundas.tableControllers.FoodService.FoodRequestTbl instance = null;
   private static TableController<Request, Integer> masterTable = null;
   /** name of table */
   protected String tbName;
@@ -28,32 +28,42 @@ public class FoodRequestTbl implements TableController<FoodRequest, ArrayList<In
   /** list that contains the objects stored in the database */
   protected ArrayList<FoodRequest> objList;
   /** relative path to the database file */
+  TableController<FoodRequest, ArrayList<Integer>> embeddedTable = null;
+
+  TableController<FoodRequest, ArrayList<Integer>> clientServerTable = null;
+
   ConnectionHandler connectionHandler = ConnectionHandler.getInstance();
 
   Connection connection = connectionHandler.getConnection();
 
+  // **
   // created constructor fo the table
   private FoodRequestTbl() throws SQLException {
+    String[] cols = {"reqID", "giftID", "quantity"};
 
-    String[] cols = {"reqID", "foodID", "quantity"};
-    colNames = Arrays.asList(cols);
     tbName = "FoodRequests";
-    pkCols = "reqID,foodID";
+    pkCols = "reqID, foodID";
+    colNames = Arrays.asList(cols);
+    objList = new ArrayList<FoodRequest>();
+    embeddedTable =
+            new FoodRequestEmbedded(tbName, colNames.toArray(new String[3]), pkCols, objList);
+    clientServerTable =
+            new FoodRequestClientServer(tbName, colNames.toArray(new String[3]), pkCols, objList);
+    connectionHandler.addTable(embeddedTable, ConnectionType.embedded);
+    connectionHandler.addTable(clientServerTable, ConnectionType.clientServer);
     masterTable = RequestTable.getInstance();
     createTable();
-
-    objList = new ArrayList<FoodRequest>();
     objList = readTable();
   }
 
   // **
   // created getInstance method for singleton implementation
-  public static FoodRequestTbl getInstance() {
+  public static edu.wpi.GoldenGandaberundas.tableControllers.FoodService.FoodRequestTbl getInstance() {
     if (instance == null) {
       synchronized (TableController.class) {
         if (instance == null) {
           try {
-            instance = new FoodRequestTbl();
+            instance = new edu.wpi.GoldenGandaberundas.tableControllers.FoodService.FoodRequestTbl();
           } catch (SQLException e) {
             e.printStackTrace();
           }
@@ -63,178 +73,42 @@ public class FoodRequestTbl implements TableController<FoodRequest, ArrayList<In
     return instance;
   }
 
+  private TableController<FoodRequest, ArrayList<Integer>> getCurrentTable() {
+    System.out.println("Connection Type: " + connectionHandler.getCurrentConnectionType());
+    switch (connectionHandler.getCurrentConnectionType()) {
+      case embedded:
+        return embeddedTable;
+      case clientServer:
+        return clientServerTable;
+      case cloud:
+        return null;
+    }
+    System.out.println(connectionHandler.getCurrentConnectionType());
+    return null;
+  }
+
   // Creates readTable method and returns an array list of Food Requests
   @Override
-  public ArrayList<FoodRequest> readTable() {
-    ArrayList tableInfo = new ArrayList<FoodRequest>(); // **
-    try {
-      PreparedStatement s = connection.prepareStatement("SElECT * FROM " + tbName + ";");
-      ResultSet r = s.executeQuery();
-      while (r.next()) {
-        tableInfo.add(new FoodRequest(masterTable.getEntry(r.getInt(1)), r.getInt(2), r.getInt(3)));
-      }
-    } catch (SQLException se) {
-      se.printStackTrace();
-      return null;
-    }
-    objList = tableInfo;
-    return tableInfo;
-  }
+  public ArrayList<FoodRequest> readTable() {return this.getCurrentTable().readTable();}
 
   @Override
-  public boolean addEntry(FoodRequest obj) {
-    if (!RequestTable.getInstance().entryExists(obj.getRequestID())) {
-      RequestTable.getInstance().addEntry(obj);
-    }
-    FoodRequest foodReq = (FoodRequest) obj; // **
-    PreparedStatement s = null;
-    try {
-      s =
-          connection.prepareStatement( // **
-              "INSERT OR IGNORE INTO " + tbName + " VALUES (?, ?, ?);");
-
-      s.setInt(1, obj.getRequestID());
-      s.setInt(2, obj.getFoodID());
-      s.setInt(3, obj.getQuantity());
-      s.executeUpdate();
-      return true;
-    } catch (SQLException e) {
-      e.printStackTrace();
-      return false;
-    }
-  }
+  public boolean addEntry(FoodRequest obj) {return this.getCurrentTable().addEntry(obj);}
 
   @Override
-  public ArrayList<FoodRequest> readBackup(String fileName) {
-    ArrayList<FoodRequest> fReqList = new ArrayList<>(); // **
-
-    try {
-      File csvFile = new File(fileName);
-      BufferedReader buffer = new BufferedReader(new FileReader(csvFile)); // reads the files
-      String currentLine = buffer.readLine(); // reads a line from the csv file
-      System.out.println(currentLine);
-      if (!currentLine
-          .toLowerCase(Locale.ROOT)
-          .trim()
-          .equals(new String("reqID,foodID,quantity"))) { // **
-        System.err.println("food request backup format not recognized"); // **
-      }
-
-      currentLine = buffer.readLine();
-
-      while (currentLine != null) { // cycles in the while loop until it reaches the end
-        String[] element = currentLine.split(","); // separates each element based on a comma
-        FoodRequest req = // **
-            new FoodRequest(
-                Integer.parseInt(element[0]),
-                element[1],
-                Integer.parseInt(element[2]),
-                Integer.parseInt(element[3]),
-                Long.parseLong(element[4]),
-                Long.parseLong(element[5]),
-                Integer.parseInt(element[6]),
-                element[8],
-                element[9],
-                Integer.parseInt(element[10]),
-                Integer.parseInt(element[11]));
-        fReqList.add(req); // adds the location to the list
-        currentLine = buffer.readLine();
-      }
-      ; // creates a Location
-
-    } catch (FileNotFoundException ex) {
-      ex.printStackTrace();
-    } catch (IOException ex) {
-      ex.printStackTrace();
-    }
-    return fReqList; // **
-  }
+  public ArrayList<FoodRequest> readBackup(String fileName) {return this.getCurrentTable().readBackup(fileName);}
 
   @Override
-  public void createTable() {
-    try {
-      PreparedStatement s =
-          connection.prepareStatement(
-              "SELECT count(*) FROM sqlite_master WHERE tbl_name = ? LIMIT 1;");
-      s.setString(1, tbName);
-      ResultSet r = s.executeQuery();
-      r.next();
-      if (r.getInt(1) != 0) {
-        return;
-      }
-    } catch (SQLException e) {
-      e.printStackTrace();
-      return;
-    }
-
-    try {
-      Class.forName("org.sqlite.JDBC");
-    } catch (ClassNotFoundException e) {
-      System.out.println("SQLite driver not found on classpath, check your gradle configuration.");
-      e.printStackTrace();
-      return;
-    }
-
-    System.out.println("SQLite driver registered!");
-
-    Statement s = null;
-    try {
-      s = connection.createStatement();
-      s.execute("PRAGMA foreign_keys = ON"); // **
-      s.execute(
-          "CREATE TABLE IF NOT EXISTS  FoodRequests("
-              + "reqID INTEGER NOT NULL, "
-              + "foodID INTEGER NOT NULL, "
-              + "quantity INTEGER NOT NULL, "
-              + "CONSTRAINT FoodRequestPK PRIMARY KEY (reqID, foodID), "
-              + "CONSTRAINT RequestFK FOREIGN KEY (reqID) REFERENCES Requests (requestID) "
-              + "ON UPDATE CASCADE "
-              + "ON DELETE CASCADE, "
-              + "CONSTRAINT FoodFK FOREIGN KEY (foodID) REFERENCES Food (foodID) "
-              + "ON UPDATE CASCADE "
-              + "ON DELETE CASCADE);");
-
-    } catch (SQLException e) {
-      e.printStackTrace();
-    }
-  }
+  public void createTable() {this.getCurrentTable().createTable();}
 
   @Override
-  public FoodRequest getEntry(ArrayList<Integer> pkID) {
-    FoodRequest foodReq = new FoodRequest();
-    if (this.entryExists(pkID)) {
-      try {
-        PreparedStatement s =
-            connection.prepareStatement(
-                "SELECT * FROM " + tbName + " WHERE (" + pkCols + ") =(?,?);");
-        s.setInt(1, pkID.get(0));
-        s.setInt(2, pkID.get(1));
-        ResultSet r = s.executeQuery();
-        r.next();
-
-        if (entryExists(pkID)) {
-          foodReq.setRequestID(r.getInt(1));
-          foodReq.setFoodID(r.getInt(2));
-          foodReq.setQuantity(r.getInt(3));
-        }
-        return foodReq; // **
-      } catch (SQLException e) {
-        e.printStackTrace();
-      }
-    }
-    return foodReq; // **
-  }
+  public FoodRequest getEntry(ArrayList<Integer> pkID) {return this.getCurrentTable().getEntry(pkID);}
 
   @Override
   public boolean loadFromArrayList(ArrayList<FoodRequest> objList) {
-    return false;
+    return this.getCurrentTable().loadFromArrayList(objList);
   }
 
-  public void writeTable() {
-    for (FoodRequest obj : objList) {
-      this.addEntry(obj);
-    }
-  }
+  public void writeTable() {this.getCurrentTable().writeTable();}
 
   /**
    * Modifies the attribute so that it is equal to value MAKE SURE YOU KNOW WHAT DATA TYPE YOU ARE
@@ -246,30 +120,8 @@ public class FoodRequestTbl implements TableController<FoodRequest, ArrayList<In
    * @return true if successful, false otherwise
    */
   // public boolean editEntry(T1 pkid, String colName, Object value)
-  public boolean editEntry(ArrayList<Integer> pkid, String colName, Object value) {
-    //    if (pkid instanceof ArrayList) {
-    //      return editEntryComposite((ArrayList<Integer>) pkid, colName, value);
-    //    }
-    try {
+  public boolean editEntry(ArrayList<Integer> pkid, String colName, Object value) {return this.getCurrentTable().editEntry(pkid, colName, value);}
 
-      PreparedStatement s =
-          connection.prepareStatement(
-              "UPDATE "
-                  + tbName
-                  + " SET "
-                  + colName
-                  + " = ? WHERE ("
-                  + colNames.get(0)
-                  + ") =(?);");
-      s.setObject(1, value);
-      s.setObject(2, pkid);
-      s.executeUpdate();
-      return true;
-    } catch (SQLException e) {
-      e.printStackTrace();
-      return false;
-    }
-  }
 
   /**
    * removes a row from the database
@@ -277,118 +129,21 @@ public class FoodRequestTbl implements TableController<FoodRequest, ArrayList<In
    * @param pkid primary key of row to be removed
    * @return true if successful, false otherwise
    */
-  public boolean deleteEntry(ArrayList<Integer> pkid) {
-    //    if (pkid instanceof ArrayList) {
-    //      return deleteEntryComposite((ArrayList<Integer>) pkid);
-    //    }
-    try {
-      PreparedStatement s =
-          connection.prepareStatement(
-              "DELETE FROM " + tbName + " WHERE " + colNames.get(0) + " = ?;");
-      s.setObject(1, pkid);
-      s.executeUpdate();
-    } catch (SQLException e) {
-      e.printStackTrace();
-    }
+  public boolean deleteEntry(ArrayList<Integer> pkid) {return this.getCurrentTable().deleteEntry(pkid);}
 
-    return false;
-  }
 
   /**
    * creates CSV file representing the objects stored in the table
    *
    * @param f filename of the to be created CSV
    */
-  public void createBackup(File f) {
-    if (objList.isEmpty()) {
-      return;
-    }
-    /* Instantiate the writer */
-    PrintWriter writer = null;
-    try {
-      writer = new PrintWriter(f);
-    } catch (FileNotFoundException e) {
-      e.printStackTrace();
-    }
-
-    /* Get the class type of the objects in the array */
-    final Class<?> type = objList.get(0).getClass();
-
-    /* Get the name of all the attributes */
-    final ArrayList<Field> classAttributes = new ArrayList<>(List.of(type.getDeclaredFields()));
-
-    boolean doesExtend = Request.class.isAssignableFrom(type);
-    if (doesExtend) {
-      final Class<?> superType = objList.get(0).getClass().getSuperclass();
-      classAttributes.addAll(0, (List.of(superType.getDeclaredFields())));
-    }
-
-    /* Write the parsed attributes to the file */
-    writer.println(classAttributes.stream().map(Field::getName).collect(Collectors.joining(",")));
-
-    /* For each object, read each attribute and append it to the file with a comma separating */
-    PrintWriter finalWriter = writer;
-    objList.forEach(
-        obj -> {
-          finalWriter.println(
-              classAttributes.stream()
-                  .map(
-                      attribute -> {
-                        attribute.setAccessible(true);
-                        String output = "";
-                        try {
-                          output = attribute.get(obj).toString();
-                        } catch (IllegalAccessException | ClassCastException e) {
-                          System.err.println("[CSVUtil] Object attribute access error.");
-                        }
-                        return output;
-                      })
-                  .collect(Collectors.joining(",")));
-          finalWriter.flush();
-        });
-    writer.close();
-  }
+  public void createBackup(File f) {this.getCurrentTable().createBackup(f);}
 
   // drop current table and enter data from CSV
-  public ArrayList<FoodRequest> loadBackup(String fileName) {
-    createTable();
-    ArrayList<FoodRequest> listObjs = readBackup(fileName);
-
-    try {
-      PreparedStatement s = connection.prepareStatement("DELETE FROM " + tbName + ";");
-      s.executeUpdate();
-      this.objList = listObjs;
-      this.writeTable();
-    } catch (SQLException e) {
-      e.printStackTrace();
-    }
-    return listObjs;
-  }
+  public ArrayList<FoodRequest> loadBackup(String fileName) {return this.getCurrentTable().loadBackup(fileName);}
 
   // checks if an entry exists
-  public boolean entryExists(ArrayList<Integer> pkID) {
-    //    if (pkID instanceof ArrayList) {
-    //      return entryExistsComposite((ArrayList<Integer>) pkID);
-    //    }
-    boolean exists = false;
-    try {
-      PreparedStatement s =
-          connection.prepareStatement(
-              "SELECT count(*) FROM " + tbName + " WHERE " + colNames.get(0) + " = ?;");
-
-      s.setObject(1, pkID);
-
-      ResultSet r = s.executeQuery();
-      r.next();
-      if (r.getInt(1) != 0) {
-        exists = true;
-      }
-
-    } catch (SQLException e) {
-      e.printStackTrace();
-    }
-    return exists;
-  }
+  public boolean entryExists(ArrayList<Integer> pkID) {return this.getCurrentTable().entryExists(pkID);}
 
   public String getTableName() {
     return tbName;
